Add vcsh hooks to use git's sparse checkout
This causes the files specified in the `pre-upgrade.00-sparse-checkout`
vcsh hook to be omitted from the actual vcsh checkout.  This permits the
vcsh managed repositories to each have files like `README.md` in the
repository and direct checkouts, yet which won't conflict when used by
vcsh.

The extra symlink `mr.post-clone.00-sparse-checkout` helps bootstrap the
sparse checkout config on an initial vcsh clone of this repository.

See the git-config(1) and git-read-tree(1) man pages for the relevant
config and sparse checkout working details.